Added linter to project.

+++ test/named_lock_test.dart
@@ -126,7 +126,7 @@ void main() {
         await Isolate.spawn(isolate_entrypoint, receiver.sendPort);
 
         // Wait for the isolate to send its message
-        return await receiver.first;
+        return await receiver.first as int;
       }
 
       String name = '${safeIntId.getId()}_named_sem';
